requirements.txt
Package               Version
--------------------- ----------
aiohttp               3.8.4
aiosignal             1.2.0
annoy                 1.17.2
async-timeout         4.0.2
asynctest             0.13.0
attrs                 22.2.0
certifi               2022.12.7
charset-normalizer    2.0.12
click                 8.0.4
cycler                0.11.0
dataclasses           0.8
datasets              2.4.0
decorator             4.4.2
diffusers             0.2.2
dill                  0.3.4
filelock              3.4.1
frozenlist            1.2.0
fsspec                2022.1.0
ftfy                  6.0.3
huggingface-hub       0.4.0
idna                  3.4
idna-ssl              1.1.0
imageio               2.15.0
importlib-metadata    4.8.3
importlib-resources   5.4.0
joblib                1.1.1
kiwisolver            1.3.1
matplotlib            3.3.4
multidict             5.2.0
multiprocess          0.70.12.2
networkx              2.5.1
nltk                  3.6.7
numpy                 1.19.5
openai-clip           1.0.1
packaging             21.3
pandas                1.1.5
Pillow                8.4.0
pip                   21.3.1
pyarrow               6.0.1
pyparsing             3.0.9
python-dateutil       2.8.2
pytz                  2022.7.1
PyWavelets            1.1.1
PyYAML                6.0
regex                 2022.10.31
requests              2.27.1
responses             0.17.0
sacremoses            0.0.53
scikit-image          0.17.2
scikit-learn          0.24.2
scipy                 1.5.4
sentence-transformers 2.2.2
sentencepiece         0.1.98
setuptools            59.6.0
six                   1.16.0
stats                 0.1.2a0
threadpoolctl         3.1.0
tifffile              2020.9.3
tokenizers            0.12.1
torch                 1.10.1
torchaudio            0.10.1
torchvision           0.11.2
tqdm                  4.64.1
transformers          4.18.0
typing_extensions     4.1.1
urllib3               1.26.14
wcwidth               0.2.6
wheel                 0.37.1
xxhash                3.2.0
yarl                  1.7.2
zipp                  3.6.0
